Dev: Updated Module page with x-editable. Dev: Updated Module entity with more fields. Dev: Added ModuleType table instead of hard coded enum. Dev: Added x-editable. Fix: Redo the Module's related table migrations.

+++ Project.Genius.Schema/Entities/DefinedTask.cs
@@ -7,7 +7,7 @@
 	{
 		#region Properties
 
-		public float Duration { get; set; }
+		public double Duration { get; set; }
 
 		[DatabaseGenerated(DatabaseGeneratedOption.Identity)]
 		public Guid Id { get; set; }
